requirements.txt
pandas==1.5.3
pytest==7.2.2
python_json_logger==2.0.7
PyYAML==6.0
